Workflow complete
The basic workflow is complete. There's a first install page that
explains what the addon does and prompts the user for next steps.
The pagemod forces all amazon transactions through smile.amazon.com
instead, and if the user hasn't configured things it will prompt them
to set it all up.

There is an addon pref that denotes if it should check for the Mozilla
Foundation, but the pref is only checked to determine if the
instructions page should load and nowhere in the code does the pref
get unchecked when setup is complete.

Next step is to split the single page mod into two, one for the smile
redirect and one for setup. Then we need some communications with the
setup content script to honor that setting and prevent it from firing
after the user has set it up.

Next after that is a watcher that constantly prompts the user if they
change the beneficiary away from the Mozilla Foundation, and a new
"never ask me about this again" pref that will prevent the watcher
from running.